refactor: CoreDataUsable 프로토콜과 익스텐션 코드 별개의 파일로 분리
